refactor: Extract component id utils from RDF resolvers
